backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cashfolio-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlmodel>=0.0.19",
    "alembic>=1.13",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=3.2,<4.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "yfinance>=0.2",
    "apscheduler>=3.10",
    "pydantic-settings>=2.2",
    "email-validator>=2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
